backend/requirements.txt
fastapi
uvicorn
requests
pydantic
pyyaml
pytest
